requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.34.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pytesseract>=0.3.10
pdf2image>=1.16.3
scikit-learn>=1.3.0
matplotlib>=3.7.0
pytest>=7.4.0
black>=23.9.0
isort>=5.12.0
ruff>=0.0.290